Create com.uber.nullaway.generics package
